read_dengue_dataset

Read the dengue dataset of [SINAVE](https://www.sinave.gob.mx/)

read_dengue_dataset(
  path,
  spatial_resolution,
  status_caso,
  des_edo_res = NULL,
  des_mpo_res = NULL
)

Arguments

path

is the path of dengue dataset.

spatial_resolution

is the spatial resolution or the administrative level. The values are country, state and municipality.

status_caso

1 probable, 2 confirmado, & 3 descartado.

des_edo_res

is a string for define the state in uppercase.

des_mpo_res

is a string for define the municipality in uppercase.

Value

a data.table object.

Author

Felipe Antonio Dzul Manzanilla felipe.dzul.m@gmail.com

Examples

read_dengue_dataset(path = "/Users/felipedzul/Dropbox/dataset/dge/denv/2025/DENGUE2_.txt", spatial_resolution = "state", des_edo_res = "NUEVO LEON", status_caso = c(1, 2))
#> Error in data.table::fread(path, header = TRUE, quote = "", select = vect_cols,     fill = TRUE, encoding = "Latin-1"): File '/Users/felipedzul/Dropbox/dataset/dge/denv/2025/DENGUE2_.txt' does not exist or is non-readable. getwd()=='/Users/fdzul/Desktop/densnv/docs/reference'